+++ apps/web/src/components/ui/cardflip.css
@@ -113,7 +113,6 @@ input[type=checkbox]:checked ~ .back {
   left: 50%;
   transform: translate(-50%, -50%);
   font-size: 5em;
-  color: transparent;
   -webkit-background-clip: text;
 }
