+++ apps/web/src/hooks/useHocuspocus.tsx
@@ -1,5 +1,5 @@
 import { HocuspocusProvider } from "@hocuspocus/provider";
-import { createContext, useContext, useState } from "react";
+import { createContext, useContext } from "react";
 import { type Doc } from "yjs";
 
 import { WS_URL } from "../lib/env";
@@ -13,7 +13,6 @@ type HocusPocusProviderProps = {
 type HocusPocusContextType = {
 	provider: HocuspocusProvider;
 	roomId: string;
-	canShow: boolean;
 	clientId?: number;
 };
 
@@ -29,17 +28,15 @@ export const useHocusPocus = () => {
 	return context;
 };
 
-export const HocusPocusProvider: React.FC<HocusPocusProviderProps> = ({
+export const HocusPocusProvider = ({
 	children,
 	roomId,
 	ydoc,
-}) => {
-	const [isLoading, setIsLoading] = useState(false);
+}: HocusPocusProviderProps) => {
 	const provider = new HocuspocusProvider({
 		url: WS_URL,
 		name: roomId,
 		document: ydoc,
-		onSynced: () => setIsLoading(true),
 		token: () => "default-token",
 		parameters: {
 			roomId,
@@ -53,7 +50,6 @@ export const HocusPocusProvider: React.FC<HocusPocusProviderProps> = ({
 			value={{
 				provider,
 				roomId,
-				canShow: isLoading,
 				clientId,
 			}}
 		>
