+++ apps/web/src/store.ts
@@ -3,11 +3,13 @@ import { devtools } from "valtio/utils";
 
 import { Issue, RoomState } from "./hooks/useRealtime";
 
-export const state = proxy<{
+interface GameState {
 	issues: Record<string, Issue[]>;
 	room: Record<string, RoomState>;
 	issuesOpen: boolean;
-}>({
+}
+
+export const state = proxy<GameState>({
 	issues: {},
 	room: {},
 	issuesOpen: false,
